requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
torch==2.3.1
h5py==3.11.0
pyyaml==6.0.1
tqdm==4.66.4
scipy==1.13.1